generacion 2 primeros casos de prueba
