+++ test/test_helper.rb
@@ -29,3 +29,4 @@ class ActiveSupport::TestCase
 
   # Add more helper methods to be used by all tests here...
 end
+
